requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# LLM providers
openai>=1.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Web framework (optional)
fastapi>=0.104.0
uvicorn>=0.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Telegram Bot
python-telegram-bot>=20.0
